python-model/requirements.txt
flask==3.1.0
flask-cors==5.0.1
torch
torchvision
pillow
requests
numpy
google-generativeai
sentence-transformers
pandas
h2>=4.0.0
hyperframe>=6.0.0
selenium>=4.0.0
clip @ git+https://github.com/openai/CLIP.git
ftfy
regex
tqdm
